# PERF: value_counts_float64 #10821