Adresses #10821 - I get a doubling of performance for `value_counts` for a `Series` with a float dtype.  Couldn't think of any new tests needed not covered by existing [here](https://github.com/pydata/pandas/blob/master/pandas/tests/test_algos.py#L416)


    In [1]: f = np.hstack([np.hstack(np.linspace(0,1,1000)  for _ in range(1000)), np.array([np.nan] * 1000)])
       ...: i = range(1000) * 1001
       ...: df = pd.DataFrame({'f':f, 'i':i})

### master

    In [2]: %timeit df['i'].value_counts()
    10 loops, best of 3: 21.1 ms per loop

    In [3]: %timeit df['f'].value_counts()
    10 loops, best of 3: 114 ms per loop


### pr

    In [3]: %timeit df['i'].value_counts()
    10 loops, best of 3: 21 ms per loop

    In [2]: %timeit df['f'].value_counts()
    10 loops, best of 3: 56 ms per loop