@shoyer 2015-08-18
looks pretty good to me...