@jreback 2015-08-18
@chris-b1 can u use fused types (with the int64) one to avoid duplicating code as much as possible
(the types are already defined near the top of the file)