@chris-b1 2015-08-18
@jreback I'm stuck on how to dispatch to the right `khash` implementation.  Starting out as below, I'm not seeing an obvious way to branch and handle both the `int64` and `float64` case?  Maybe I'm missing something.  Thanks.

    cpdef value_counts_numeric(sixty_four_bit_scalar[:] values):
        ....
        if sixty_four_bit_scale is float64_t:
            table = ??