@jreback 2015-08-18
@chris-b1 yes that is right. see the ``append_data`` function. This is an odd looking cython syntax for fused type specialization, but apparetnly is a compile type thing. see docs [here](http://docs.cython.org/src/userguide/fusedtypes.html#type-checking-specializations)