@chris-b1 2015-08-18
@jreback  Sorry, my question wasn't clear - I (somewhat) see how the fused types work, but I'm specifically stuck on accessing the right `khash` functions given a fused type.  Below are things I tried that don't work:


No `cdef` for `table` - cython thinks it's a python object

    cpdef value_counts_numeric(sixty_four_bit_scalar[:] values):
        ....
        if sixty_four_bit_scale is float64_t:
            table = kh_init_float64()

Conditional `cdef`, which aren't valid.

    cpdef value_counts_numeric(sixty_four_bit_scalar[:] values):
        ....
        if sixty_four_bit_scale is float64_t:
            cdef table *kh_float64
            table = kh_init_float64()


Fused khash type - invalid to use within a `cdef`

    ctypedef fused kh_numeric:
        kh_float64_t
        kh_int64_t

    cpdef value_counts_numeric(sixty_four_bit_scalar[:] values):
        cdef table * kh_numeric
        if sixty_four_bit_scale is float64_t:
            cdef table *kh_float64
            table = kh_init_float64()