@jreback 2015-08-18
```
ctypedef fused kh_numeric:
    kh_float64_t
    kh_int64_t

cpdef value_counts_numeric(sixty_four_bit_scalar[:] values):
    cdef table * kh_numeric
    if sixty_four_bit_scale is float64_t:
        table = kh_init_float64()
```
doesn't this work?