@chris-b1 2015-08-18
Doesn't seem to.  Cython actually crashes so the message isn't super helpful (see below), but I figured it was invalid?

    Error compiling Cython file:
    ------------------------------------------------------------
    ...
    cpdef value_counts_numeric(sixty_four_bit_scalar[:] values, bint dropna = 0):
        cdef:
            Py_ssize_t i
            sixty_four_bit_scalar[:] result_keys
            int64_t[:] result_counts
            kh_numeric *table
        ^
    ------------------------------------------------------------

    pandas\hashtable.pyx:909:8: Compiler crash in AnalyseDeclarationsTransform

    ModuleNode.body = StatListNode(hashtable.pyx:3:0)
    StatListNode.stats[38] = StatListNode(hashtable.pyx:904:6)
    StatListNode.stats[0] = CompilerDirectivesNode(hashtable.pyx:904:6)
    CompilerDirectivesNode.body = StatListNode(hashtable.pyx:904:6)
    StatListNode.stats[0] = CFuncDefNode(hashtable.pyx:904:6,
        args = [...]/2,
        decorators = [...]/0,
        has_fused_arguments = True,
        modifiers = [...]/0,
        overridable = 1,
        visibility = u'private')
    FusedCFuncDefNode.stats[0] = CFuncDefNode(hashtable.pyx:904:6,
        args = [...]/2,
        decorators = [...]/0,
        modifiers = [...]/0,
        overridable = 1,
        visibility = u'private')
    File 'Nodes.py', line 416, in analyse_declarations: StatListNode(hashtable.pyx:905:4,
        is_terminator = True)
    File 'Nodes.py', line 416, in analyse_declarations: StatListNode(hashtable.pyx:906:8)
    File 'Nodes.py', line 1282, in analyse_declarations: CVarDefNode(hashtable.pyx:909:8,
        modifiers = [...]/0,
        visibility = u'private')
    File 'Nodes.py', line 990, in analyse: CSimpleBaseTypeNode(hashtable.pyx:909:8,
        module_path = [...]/0,
        name = u'kh_numeric',
        signed = 1)

    Compiler crash traceback from this point on:
    File "C:\Users\Chris\Anaconda\envs\pandas-dev\lib\site-packages\Cython\Compiler\Nodes.py", line 990, in analyse
        type = scope.lookup_type(self.name)
    File "C:\Users\Chris\Anaconda\envs\pandas-dev\lib\site-packages\Cython\Compiler\Symtab.py", line 810, in lookup_type
        return entry.type.specialize(self.fused_to_specific)
    File "C:\Users\Chris\Anaconda\envs\pandas-dev\lib\site-packages\Cython\Compiler\PyrexTypes.py", line 1357, in specialize
        return values[self]
    KeyError: FusedType(name=u'kh_numeric')