@jreback 2015-08-18
@chris-b1 hmm, not real familiar why you can't do this. Let me have a couple of people look at it. If you make progress, lmk.