@jreback 2015-08-18
cc @insertinterestingnamehere any idea how to specify fused types?