@kawochen 2015-08-18
```
cdef:
    kh_int64_t * table_int
    kh_float64_t * table_float
```
`cdef` both, then use the appropriate one.  If you don't to `cdef` both, then include `kh_numeric` in the signature as a dummy arg and use indexing to dispatch from the one argument version.