@insertinterestingnamehere 2015-08-18
@jreback I just tried doing this myself and see the same error. The syntax @chris-b1 used is correct, but Cython isn't handling it correctly. I'm looking for a workaround now.