@insertinterestingnamehere 2015-08-18
Okay, on second thought, the problem is that Cython can't determine which version of the fused khash type to use when the pointer is declared. The approach suggested by @kawochen is probably the simplest. I'll see if there's a nice way to tie it all together.