@kawochen 2015-08-18
Yes. You could make it visible in the signature and then dispatch to the right version, but I don't know if saving that one pointer is worth going through the hoops.