@jreback 2015-08-18
yeh, so if this turns out to be more hoops than worth it....then can just revert.