@kawochen 2015-08-18
@jreback  I was just comparing the two approaches I suggested.  The first one seems effective enough -- just one extra pointer.