@insertinterestingnamehere 2015-08-18
Okay, fused types aren't terribly effective when there are so many type-specific function calls in the core part of each function, but I was able to simplify things a little. I'm chasing down some test failures now, but you can get the main idea of what this would look like from what I have thus far. https://github.com/insertinterestingnamehere/pandas/commit/009f4df01cadc32fe7a6b3809174aa81735b7f05 shows how to combine the `value_count` routines, and https://github.com/insertinterestingnamehere/pandas/commit/f28a443f7ead3cfd08ca857294fe289cf17f1c4a shows how to do that for the `build_count_table` routines. To get the latter building properly, a bunch of casts to and from void pointers were necessary since there isn't currently a good way to make fused types depend on one another. I'd recommend only using fused types for the `val_count` functions since it makes the code more concise without complicating things too much.