@insertinterestingnamehere 2015-08-18
Alright, https://github.com/insertinterestingnamehere/pandas/commit/8ae9b9af259e3b1f6ab1f56b3d43abc04535fe33 and https://github.com/insertinterestingnamehere/pandas/commit/b9b54ae00eaa4278e0949f4df7da7cbf4aaedd34 should do nicely. Feel free to cherry pick the first or both of them as needed.