@insertinterestingnamehere 2015-08-18
Another idea just occurred to me. The type casts can be avoided by allowing the distinct fused types to expand fully and then throwing an error if the wrong combination is used. https://github.com/insertinterestingnamehere/pandas/commit/ba85304acff1f3096056e64e819be3ef3c4483cf does that. That makes it all a lot cleaner. Would you like me to make a PR against the branch used here?